% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lazy.file.end.R
\name{lazy.file.end}
\alias{lazy.file.end}
\title{End LaTeX Documents}
\usage{
lazy.file.end()
}
\description{
Provides the code to end a LaTeX document
}
\examples{
lazy.file.end()

}
\author{
Benjamin Nutter
}
