% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lazy.landscape.R
\name{lazy.landscape}
\alias{lazy.landscape}
\title{Landscape Page Orientation}
\usage{
lazy.landscape(begin = TRUE)
}
\arguments{
\item{begin}{If \code{TRUE}, begins the landscape environment.
Otherwise, the environment is closed}
}
\value{
Returns a string that either begins or ends a 
 landscape environment
}
\description{
Allows the user to turn on or off landscape page orientation
}
\details{
This function has no effect on HTML or markdown files.  
  The page orientation is determined by the browser
}
\examples{
lazy.landscape()
lazy.landscape(begin=FALSE)
}
\author{
Benjamin Nutter
}
