% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lazy.label.R, R/lazy.ref.R
\name{lazy.label}
\alias{lazy.label}
\alias{lazy.ref}
\title{Reference Tables, Figures, Sections, and Pages}
\usage{
lazy.label(label)

lazy.ref(label, text, page = FALSE, link = TRUE)
}
\arguments{
\item{label}{A character(1) giving the name of the to be 
created or referenced}

\item{text}{For HTML, the text to be hyperlinked for the reference.
If missing, this is set to \code{"(link)"}}

\item{page}{Indicates if the page number on which the label 
lies should be returned or the object number.  This only applies 
to LaTeX files}

\item{link}{for LaTeX files, should the reference link to the object}
}
\description{
Provides the code to label and reference objects that 
  have a label statement with them
}
\examples{
lazy.label("Label1")
lazy.ref("Label1")
lazy.ref("Label1", page=TRUE)

}
\author{
Benjamin Nutter
}
