\name{Scout}
\alias{Scout}
\title{Scout Advancement Data}
\description{Advancement Data for 105 Scouts in the Midwest.

  These are quasi-fictional data taken from a midwestern troop.  Names have 
  been randomly assigned from a list of popular baby names on the internet.
  Dates for each subject have been altered by subtracting a randomly 
  selected number within a particular range.  Otherwise, the data are real.
  
  In Boy Scouts, each boy may progress through a number of ranks.  These are,
  in ascending order, Scout, Tenderfoot, Second Class, First Class, Star,
  Life, and Eagle.
}

\usage{Scout}

\format{A data frame with 105 rows and 26 variables.  Variables include
  date of birth and date of each rank advancement.  There are also variables
  for the number of days between earning Scout and each successive rank.}

\source{Midwestern Scout Troop}

\keyword{ datasets }
