% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lazy.citation.R
\name{lazy.citation}
\alias{lazy.citation}
\title{Add R or R Package Citation}
\usage{
lazy.citation(
  pkg = NULL,
  author = TRUE,
  title = TRUE,
  org = TRUE,
  address = TRUE,
  volume = TRUE,
  year = TRUE,
  note = TRUE
)
}
\arguments{
\item{pkg}{a character(1) vector giving the name of a package.  If \code{NULL}, a citation for R is produced.}

\item{author}{Include author name}

\item{title}{Include title of package}

\item{org}{Include organization name}

\item{address}{Include address}

\item{volume}{Include volume}

\item{year}{include year of publication}

\item{note}{include the note on the citation.}
}
\description{
Generates code for the citation of R or an R package.
}
\details{
Not every option is populated in every package.  
Future improvements might include automatic detection of NULL fields, 
but for now, observing the output with all the options set to \code{TRUE}
will tell you which ones are empty.
}
\examples{
lazy.citation()
lazy.citation(pkg="lazyWeave", org=FALSE, address=FALSE, volume=FALSE)
}
\author{
Benjamin Nutter
}
