\name{lazy.file.end}
\alias{lazy.file.end}
\title{End LaTeX Documents}
\description{Provides the code to end a LaTeX document.}
\usage{
lazy.file.end()
}

\value{ Returns the string "\\n\\end{document}"}

\author{Benjamin Nutter \email{nutterb@ccf.org}}

\examples{
lazy.file.end()
}

\keyword{ methods }

