\name{lazy.verbatim}
\alias{lazy.verbatim.start}
\alias{lazy.verbatim.end}
\title{Verbatim Environments}
\description{Text within a verbatim block appears exactly as type, including whitespace.
  This is useful when inserting code into a document.}
\usage{
lazy.verbatim.start()
lazy.verbatim.end()
}


\details{
  A verbatim block takes any text entered and typsets it exactly as it was entered.  
  White space is preserved and the font changes.  This is typically done to display
  code, since the whitespace may preserve readability.
}

\author{Benjamin Nutter \email{nutterb@ccf.org}}


\keyword{methods}

